Add x and y properties to dxp_socket_desktop
